Move most reductions up to the Frame level.
